id_set ut re-organized
