+++ demisto_sdk/commands/common/tests/update_id_set_test.py
@@ -8,13 +8,14 @@ from tempfile import mkdtemp
 
 import pytest
 from demisto_sdk.commands.common.git_tools import git_path
-from demisto_sdk.commands.common.update_id_set import (find_duplicates,
-                                                       get_integration_data,
-                                                       get_playbook_data,
-                                                       get_script_data,
-                                                       has_duplicate,
-                                                       re_create_id_set)
+from demisto_sdk.commands.common.update_id_set import (
+    find_duplicates, get_classifier_data, get_fields_by_script_argument,
+    get_incident_field_data, get_incident_fields_by_playbook_input,
+    get_incident_type_data, get_indicator_type_data, get_integration_data,
+    get_layout_data, get_mapper_data, get_playbook_data, get_script_data,
+    get_values_for_keys_recursively, has_duplicate, re_create_id_set)
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
+from TestSuite.utils import IsEqualFunctions
 
 TESTS_DIR = f'{git_path()}/demisto_sdk/tests'
 
@@ -238,12 +239,10 @@ class TestIntegrations:
         }
     }
 
-    # @pytest.mark.skip
     def test_get_integration_data(self):
         """
         Test for getting all the integration data
         """
-        # mocker.patch.object('get_pack_name', return_value='DummyPack')
         non_unified_file_path = TESTS_DIR + '/test_files/DummyPack/Integrations/DummyIntegration/DummyIntegration.yml'
         unified_file_path = TESTS_DIR + '/test_files/DummyPack/Integrations/integration-DummyIntegration.yml'
 
@@ -259,16 +258,573 @@ class TestIntegrations:
             returned = pair[0]
             constant = pair[1]
 
-            assert list(returned.keys()) == list(constant.keys())
+            assert IsEqualFunctions.is_lists_equal(list(returned.keys()), list(constant.keys()))
+
             const_data = constant.get('Dummy Integration')
             returned_data = returned.get('Dummy Integration')
 
-            for key, value in returned_data.items():
-                assert key in const_data.keys()
-                if isinstance(value, list):
-                    assert sorted(value) == sorted(const_data.get('key'))
-                else:
-                    assert value == const_data.get(key)
+            assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
+
+
+class TestScripts:
+    SCRIPT_DATA = {
+        "DummyScript": {
+            "name": "DummyScript",
+            "file_path": TESTS_DIR + "/test_files/DummyPack/Scripts/DummyScript.yml",
+            "fromversion": "5.0.0",
+            "tests": [
+                "No test - no need to test widget"
+            ]
+        }
+    }
+
+    @staticmethod
+    def test_get_script_data():
+        """
+        Test for getting the script data
+        """
+        file_path = TESTS_DIR + '/test_files/DummyPack/Scripts/DummyScript.yml'
+        data = get_script_data(file_path)
+
+        assert IsEqualFunctions.is_lists_equal(list(data.keys()), list(TestScripts.SCRIPT_DATA.keys()))
+
+        const_data = TestScripts.SCRIPT_DATA.get('DummyScript')
+        returned_data = data.get('DummyScript')
+
+        assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
+
+
+class TestPlaybooks:
+    PLAYBOOK_DATA = {
+        "name": "Dummy Playbook",
+        "file_path": TESTS_DIR + "/test_files/DummyPack/Playbooks/DummyPlaybook.yml",
+        "fromversion": "4.5.0",
+        "implementing_scripts": [
+            "XDRSyncScript",
+            "StopScheduledTask",
+        ],
+        "implementing_playbooks": [
+            "Calculate Severity - Standard",
+            "Palo Alto Networks - Malware Remediation",
+        ],
+        "command_to_integration": {
+            "xdr-update-incident": "",
+            "autofocus-sample-analysis": ""
+        },
+        "tests": [
+            "No Test"
+        ],
+        "skippable_tasks": [
+            "StopScheduledTask",
+            "autofocus-sample-analysis"
+        ]
+    }
+
+    @staticmethod
+    def test_get_playbook_data():
+        """
+        Test for getting the playbook data
+        """
+        file_path = TESTS_DIR + '/test_files/DummyPack/Playbooks/DummyPlaybook.yml'
+        data = get_playbook_data(file_path)['Dummy Playbook']
+
+        assert IsEqualFunctions.is_dicts_equal(data, TestPlaybooks.PLAYBOOK_DATA)
+
+    @staticmethod
+    def test_get_playbook_data_2():
+        """
+        Given
+            - A playbook file called playbook-with-incident-fields.yml
+
+        When
+            - parsing playbook files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-with-incident-fields.yml'
+        result = get_playbook_data(test_dir)
+        result = result.get('Arcsight - Get events related to the Case')
+        assert 'name' in result.keys()
+        assert 'file_path' in result.keys()
+        assert 'implementing_scripts' in result.keys()
+        assert 'implementing_scripts' in result.keys()
+        assert 'command_to_integration' in result.keys()
+        assert 'tests' in result.keys()
+        assert 'incident_fields' in result.keys()
+        assert 'indicator_fields' in result.keys()
+
+    @staticmethod
+    def test_get_playbook_data_no_fields():
+        """
+        Given
+            - A playbook file called playbook-no-incident-fields.yml without any
+                incident or indicator fields that it depends on.
+
+        When
+            - parsing playbook files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-no-incident-fields.yml'
+        result = get_playbook_data(test_dir)
+        result = result.get('Arcsight - Get events related to the Case')
+        assert 'name' in result.keys()
+        assert 'file_path' in result.keys()
+        assert 'implementing_scripts' in result.keys()
+        assert 'implementing_scripts' in result.keys()
+        assert 'command_to_integration' in result.keys()
+        assert 'tests' in result.keys()
+        assert 'incident_fields' not in result.keys()
+        assert 'indicator_fields' not in result.keys()
+
+
+class TestLayouts:
+    @staticmethod
+    def test_get_layout_data():
+        """
+        Given
+            - A layout file called layout-to-test.json
+
+        When
+            - parsing layout files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/layout-to-test.json'
+        result = get_layout_data(test_dir)
+        result = result.get('urlRep')
+        assert 'kind' in result.keys()
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'toversion' in result.keys()
+        assert 'path' in result.keys()
+        assert 'typeID' in result.keys()
+        assert 'incident_and_indicator_types' in result.keys()
+        assert 'incident_and_indicator_fields' in result.keys()
+
+    @staticmethod
+    def test_get_layout_data_no_incident_types_and_fields():
+        """
+        Given
+            - A layout file called layout-to-test.json that doesnt have related incident fields and indicator fields
+
+        When
+            - parsing layout files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/layout-to-test-no-types-fields.json'
+        result = get_layout_data(test_dir)
+        result = result.get('urlRep')
+        assert 'kind' in result.keys()
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'toversion' in result.keys()
+        assert 'path' in result.keys()
+        assert 'typeID' in result.keys()
+        assert 'incident_and_indicator_types' in result.keys()
+        assert 'incident_and_indicator_fields' not in result.keys()
+
+
+class TestIncidentFields:
+    @staticmethod
+    def test_get_incident_fields_data():
+        """
+        Given
+            - An incident field file called incidentfield-to-test.json
+
+        When
+            - parsing incident field files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidentfield-to-test.json'
+        result = get_incident_field_data(test_dir, [])
+        result = result.get('incidentfield-test')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'toversion' in result.keys()
+        assert 'incident_types' in result.keys()
+        assert 'scripts' in result.keys()
+
+    @staticmethod
+    def test_get_incident_fields_data_no_types_scripts():
+        """
+        Given
+            - An incident field file called incidentfield-to-test-no-types_scripts.json with no script or incident type
+            related to it
+
+        When
+            - parsing incident field files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidentfield-to-test-no-types_scripts.json'
+        result = get_incident_field_data(test_dir, [])
+        result = result.get('incidentfield-test')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'toversion' in result.keys()
+        assert 'incident_types' not in result.keys()
+        assert 'scripts' not in result.keys()
+
+
+class TestIndicatorType:
+    @staticmethod
+    def test_get_indicator_type_data():
+        """
+        Given
+            - An indicator type file called reputation-indicatortype.json.
+
+        When
+            - parsing indicator type files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/reputation-indicatortype.json'
+
+        result = get_indicator_type_data(test_dir)
+        result = result.get('indicator-type-dummy')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'integrations' in result.keys()
+        assert 'scripts' in result.keys()
+        assert "dummy-script" in result.get('scripts')
+        assert "dummy-script-2" in result.get('scripts')
+        assert "dummy-script-3" in result.get('scripts')
+
+    @staticmethod
+    def test_get_indicator_type_data_no_integration_no_scripts():
+        """
+        Given
+            - An indicator type file called reputation-indicatortype_no_script_no_integration.json without any
+                integrations or scripts that it depends on.
+
+        When
+            - parsing indicator type files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/' \
+                   f'reputation-indicatortype_no_script_no_integration.json'
+
+        result = get_indicator_type_data(test_dir)
+        result = result.get('indicator-type-dummy')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'integrations' not in result.keys()
+        assert 'scripts' not in result.keys()
+
+
+class TestIncidentTypes:
+    @staticmethod
+    def test_get_incident_types_data():
+        """
+        Given
+            - An incident type file called incidenttype-to-test.json
+
+        When
+            - parsing incident type files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidenttype-to-test.json'
+        result = get_incident_type_data(test_dir)
+        result = result.get('dummy incident type')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'playbooks' in result.keys()
+        assert 'scripts' in result.keys()
+
+    @staticmethod
+    def test_get_incident_types_data_no_playbooks_scripts():
+        """
+        Given
+            - An incident type file called incidenttype-to-test-no-playbook-script.json with no script or playbook
+            related to it
+
+        When
+            - parsing incident type files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidenttype-to-test-no-playbook-script.json'
+
+        result = get_incident_type_data(test_dir)
+        result = result.get('dummy incident type')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'playbooks' not in result.keys()
+        assert 'scripts' not in result.keys()
+
+
+class TestClassifiers:
+    @staticmethod
+    def test_get_classifiers_data():
+        """
+        Given
+            - A classifier file called classifier-to-test.json
+
+        When
+            - parsing classifier files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-to-test.json'
+        result = get_classifier_data(test_dir)
+        result = result.get('dummy classifier')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'incident_types' in result.keys()
+        assert 'dummy incident type' in result['incident_types']
+        assert 'dummy incident type 2' in result['incident_types']
+        assert 'dummy incident type 3' in result['incident_types']
+
+    @staticmethod
+    def test_get_classifiers_data_no_types_scripts():
+        """
+        Given
+            - An classifier file called classifier-to-test-no-incidenttypes.json with incident type
+            related to it
+
+        When
+            - parsing classifier files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-to-test-no-incidenttypes.json'
+
+        result = get_classifier_data(test_dir)
+        result = result.get('dummy classifier')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'incident_types' not in result.keys()
+        assert 'incident_fields' not in result.keys()
+
+
+class TestMappers:
+    @staticmethod
+    def test_get_mappers_data():
+        """
+        Given
+            - A mapper file called classifier-mapper-to-test.json
+
+        When
+            - parsing mapper files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-mapper-to-test.json'
+        result = get_mapper_data(test_dir)
+        result = result.get('dummy mapper')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'incident_types' in result.keys()
+        assert 'incident_fields' in result.keys()
+        assert 'dummy incident type' in result['incident_types']
+        assert 'dummy incident type 1' in result['incident_types']
+        assert 'dummy incident type 2' in result['incident_types']
+        assert 'dummy incident field' in result['incident_fields']
+        assert 'dummy incident field 1' in result['incident_fields']
+        assert 'dummy incident field 2' in result['incident_fields']
+        assert 'dummy incident field 3' in result['incident_fields']
+
+    @staticmethod
+    def test_get_mappers_data_no_types_fields():
+        """
+        Given
+            - An mapper file called classifier-mapper-to-test-no-types-fields.json with incident type
+            related to it
+
+        When
+            - parsing mapper files
+
+        Then
+            - parsing all the data from file successfully
+        """
+        test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-mapper-to-test-no-types-fields.json'
+
+        result = get_mapper_data(test_dir)
+        result = result.get('dummy mapper')
+        assert 'name' in result.keys()
+        assert 'fromversion' in result.keys()
+        assert 'incident_types' not in result.keys()
+        assert 'incident_fields' not in result.keys()
+
+
+class TestGenericFunctions:
+    @staticmethod
+    def test_get_values_for_keys_recursively():
+        """
+        Given
+            - A list of keys to extract their values from a dict
+
+        When
+            - Extracting data from nested elements in the json
+
+        Then
+            - Extracting the values from all the levels of nesting in the json
+        """
+
+        test_dict = {
+            'id': 1,
+            'nested': {
+                'x1': 1,
+                'x2': 'x2',
+                'x3': False,
+                'x4': [
+                    {
+                        'x1': 2,
+                        'x2': 55
+                    },
+                    {
+                        'x3': 1,
+                        'x2': True
+                    }
+                ]
+            },
+            'x2': 4.0
+        }
+
+        test_keys = ['x1', 'x2', 'x3']
+
+        expected = {
+            'x1': [1, 2],
+            'x2': ['x2', 55, True, 4.0],
+            'x3': [False, 1]
+        }
+
+        assert expected == get_values_for_keys_recursively(test_dict, test_keys)
+
+    INPUT_WITH_INCIDENT_FIELD_SIMPLE = {
+        "key": "AlertID",
+        "value": {
+            "simple": "${incident.field_name}"
+        },
+        "required": False
+    }
+
+    INPUT_WITH_INCIDENT_FIELD_COMPLEX1 = {
+        "key": "AlertID",
+        "value": {
+            "complex": {
+                "root": "incident",
+                "accessor": "field_name"
+            }
+        },
+        "required": False
+    }
+
+    INPUT_WITH_INCIDENT_FIELD_COMPLEX2 = {
+        "key": "AlertID",
+        "value": {
+            "complex": {
+                "root": "incident.field_name",
+                "accessor": "username"
+            }
+        },
+        "required": False
+    }
+
+    INPUT_SIMPLE_WITHOUT_INCIDENT_FIELD = {
+        "key": "AlertID",
+        "value": {
+            "simple": "${not_incident.field_name}"
+        },
+        "required": False
+    }
+
+    INPUT_COMPLEX_WITHOUT_INCIDENT_FIELD = {
+        "key": "AlertID",
+        "value": {
+            "complex": {
+                "root": "something",
+                "accessor": "username"
+            }
+        },
+        "required": False
+    }
+
+    INPUTS = [
+        (INPUT_WITH_INCIDENT_FIELD_SIMPLE, True),
+        (INPUT_WITH_INCIDENT_FIELD_COMPLEX1, True),
+        (INPUT_WITH_INCIDENT_FIELD_COMPLEX2, True),
+        (INPUT_SIMPLE_WITHOUT_INCIDENT_FIELD, False),
+        (INPUT_COMPLEX_WITHOUT_INCIDENT_FIELD, False)
+    ]
+
+    @staticmethod
+    @pytest.mark.parametrize('playbook_input, are_there_incident_fields', INPUTS)
+    def test_get_incident_fields_by_playbook_input(playbook_input, are_there_incident_fields):
+        """
+        Given
+            - A list of playbook inputs
+
+        When
+            - Searching for dependent incident fields
+
+        Then
+            -  Finding all dependent incident fields in the input
+        """
+
+        result = get_incident_fields_by_playbook_input(input=playbook_input.get('value'))
+        if are_there_incident_fields:
+            assert "field_name" in result
+        else:
+            assert result == set()
+
+    EXAMPLE_TASK_WITH_SIMPLE_SCRIPT_ARGUMENTS = {
+        "id": "ID",
+        "scriptarguments": {
+            "field_name": {
+                "simple": "${inputs.IndicatorTagName}"
+            }
+        }
+    }
+
+    EXAMPLE_TASK_WITH_CUSTOM_FIELDS_SCRIPT_ARGUMENTS = {
+        "id": "ID",
+        "scriptarguments": {
+            "customFields": {
+                "simple": '[{"field_name":"${inputs.IndicatorTagName}"}]'
+            }
+        }
+    }
+
+    TASK_INPUTS = [
+        # EXAMPLE_TASK_WITH_SIMPLE_SCRIPT_ARGUMENTS,
+        EXAMPLE_TASK_WITH_CUSTOM_FIELDS_SCRIPT_ARGUMENTS
+    ]
+
+    @staticmethod
+    @pytest.mark.parametrize('task', TASK_INPUTS)
+    def test_get_fields_by_script_argument(task):
+        """
+        Given
+            - A list of playbook tasks
+
+        When
+            - Searching for dependent incident fields in the task script arguments
+
+        Then
+            - Finding all dependent incident fields in the task
+        """
+
+        result = get_fields_by_script_argument(task)
+        assert "field_name" in result
 
 
 class TestFlow(unittest.TestCase):
@@ -338,66 +894,6 @@ class TestFlow(unittest.TestCase):
         "name": "Account ID",
         "fromVersion": "5.0.0"
     }
-
-    SCRIPT_DATA = {
-        "DummyScript": {
-            "name": "DummyScript",
-            "file_path": TESTS_DIR + "/test_files/DummyPack/Scripts/DummyScript.yml",
-            "fromversion": "5.0.0",
-            "tests": [
-                "No test - no need to test widget"
-            ]
-        }
-    }
-
-    PLAYBOOK_DATA = {
-        "name": "Dummy Playbook",
-        "file_path": TESTS_DIR + "/test_files/DummyPack/Playbooks/DummyPlaybook.yml",
-        "fromversion": "4.5.0",
-        "implementing_scripts": [
-            "XDRSyncScript",
-            "StopScheduledTask",
-        ],
-        "implementing_playbooks": [
-            "Calculate Severity - Standard",
-            "Palo Alto Networks - Malware Remediation",
-        ],
-        "command_to_integration": {
-            "xdr-update-incident": "",
-            "autofocus-sample-analysis": ""
-        },
-        "tests": [
-            "No Test"
-        ],
-        'pack': 'DummyPack',
-        "skippable_tasks": [
-            "StopScheduledTask",
-            "autofocus-sample-analysis"
-        ]
-    }
-
-    def test_get_script_data(self):
-        """
-        Test for getting the script data
-        """
-        file_path = TESTS_DIR + '/test_files/DummyPack/Scripts/DummyScript.yml'
-        data = get_script_data(file_path)
-        self.assertDictEqual(data, TestFlow.SCRIPT_DATA)
-
-    def test_get_playbook_data(self):
-        """
-        Test for getting the playbook data
-        """
-        file_path = TESTS_DIR + '/test_files/DummyPack/Playbooks/DummyPlaybook.yml'
-        data = get_playbook_data(file_path)['Dummy Playbook']
-        self.assertEqual(data['name'], TestFlow.PLAYBOOK_DATA['name'])
-        self.assertEqual(data['file_path'], TestFlow.PLAYBOOK_DATA['file_path'])
-        self.assertEqual(data['fromversion'], TestFlow.PLAYBOOK_DATA['fromversion'])
-        self.assertListEqual(data['tests'], TestFlow.PLAYBOOK_DATA['tests'])
-        self.assertListEqual(data['skippable_tasks'], TestFlow.PLAYBOOK_DATA['skippable_tasks'])
-        self.assertSetEqual(set(data['implementing_playbooks']), set(TestFlow.PLAYBOOK_DATA['implementing_playbooks']))
-        self.assertListEqual(data['tests'], TestFlow.PLAYBOOK_DATA['tests'])
-        self.assertDictEqual(data['command_to_integration'], TestFlow.PLAYBOOK_DATA['command_to_integration'])
 
     # TODO: unskip
     @pytest.mark.skip
@@ -496,9 +992,3 @@ class TestFlow(unittest.TestCase):
             assert any('temp-report-dup-check' in i for i in dup_data)
             assert any('temp-widget-dup-check' in i for i in dup_data)
             assert any('dup-check-dashbaord' in i for i in dup_data)
-            assert any('layout-dup-check-id' in i for i in dup_data)
-            assert any('incident_account_field_dup_check' in i for i in dup_data)
-
-
-if __name__ == '__main__':
-    unittest.main()
